+++ frontend/src/views/AdminView.vue
@@ -66,7 +66,7 @@ import { ref } from 'vue'
 import DashboardLayout from '../components/admin/DashboardLayout.vue'
 import SidebarItem from '../components/admin/SidebarItem.vue'
 import DualInteractiveBoxes from '../components/admin/DualInteractiveBoxes.vue'
-import EventActionsBox from '../components/EventActionsBox.vue'
+import EventActionsBox from "../components/admin/eventview/EventActionsBox.vue"
 import EventCalendarBox from '../components/EventCalendarBox.vue'
 import EventTable from "../components/admin/eventview/EventTable.vue";
 import { h } from 'vue'
